+++ src/constants.ts
@@ -1,12 +1,14 @@
+export const REQUESTSPERSECOND = 10
+export const TOKEN_URL_PATH = '/oauth/token'
+export const PROCESSINGWAIT = 60 * 1000
+export const RETRIES = 5
 export const WORKFLOW_NAME = 'Email Sender'
 export const PADDING = '   '
 export const MSDAY = 86400000
 export const IDENTITYNOTFOUNDWAIT = 5000
 export const TASKRESULTWAIT = 1000
-export const PROCESSINGWAIT = 60 * 1000
 export const IDENTITYNOTFOUNDRETRIES = 5
 export const TASKRESULTRETRIES = 5
-export const REQUESTSPERSECOND = 10
 export const SEPARATOR = ' | '
 export const UNIQUEFORMNAME = 'Identity Merging'
 export const EDITFORMNAME = 'Fusion account edit form'
